# docs/openapi.yaml
openapi: "3.0.3"
info:
  title: ZuzuFlow API
  version: "1.0.0"
  description: |
    REST API for ZuzuFlow.
    All endpoints are under `/api`. Executions broadcast real-time events
    over WebSocket at `ws://<host>/ws/executions?id=<executionId>`.
servers:
  - url: http://localhost:4000/api
    description: Local development
...
components:
  schemas:
...
    ExecutionStatus:
      type: string
      enum: [pending, running, completed, failed, cancelled, timed_out]
...
    Execution:
      type: object
      properties:
        id:
          type: string
          format: uuid
        workflowId:
          type: string
          format: uuid
        status:
          $ref: "#/components/schemas/ExecutionStatus"
        triggerPayload:
          type: object
          nullable: true
        output:
          type: object
          nullable: true
        error:
          type: string
          nullable: true
        startedAt:
          type: string
          format: date-time
        completedAt:
          type: string
          format: date-time
          nullable: true
    ExecutionList:
      type: object
      properties:
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer
        items:
          type: array
          items:
            $ref: "#/components/schemas/Execution"
    CredentialKind:
      type: string
      enum:
        - postgres
        - smtp
        - sendgrid
        - mqtt
        - http_bearer
        - http_basic
        - http_api_key
        - generic
    Credential:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          example: "Production Postgres"
        kind:
          $ref: "#/components/schemas/CredentialKind"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
paths:
...
  /executions/start:
    post:
      tags: [Executions]
      summary: Start a workflow execution
      description: |
        Manually triggers a workflow. The workflow must be in `active` status
        (save + activate it first). Returns immediately with the new execution
        record — monitor progress via WebSocket.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [workflowId]
              properties:
                workflowId:
                  type: string
                  format: uuid
                triggerPayload:
                  type: object
                  description: Arbitrary JSON passed as the trigger node's output
      responses:
        "201":
          description: Execution started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Execution"
        "409":
          description: Workflow is not active
  /executions:
    get:
      tags: [Executions]
      summary: List executions
      parameters:
        - name: workflowId
          in: query
          schema:
            type: string
            format: uuid
          description: Filter by workflow
        - name: status
          in: query
          schema:
            $ref: "#/components/schemas/ExecutionStatus"
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: Paginated execution list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExecutionList"
  /executions/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid

    get:
      tags: [Executions]
      summary: Get an execution (includes node logs)
      responses:
        "200":
          description: Execution detail with logs
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Execution"
                  - type: object
                    properties:
                      logs:
                        type: array
                        items:
                          type: object
                          properties:
                            id: { type: string }
                            nodeId: { type: string }
                            nodeKind: { type: string }
                            level: { type: string, enum: [debug, info, warn, error] }
                            message: { type: string }
                            data: { type: object }
                            createdAt: { type: string, format: date-time }

    delete:
      tags: [Executions]
      summary: Cancel or delete an execution
      description: |
        If the execution is still running, it is cancelled in Temporal and marked `cancelled`.
        If the execution is already in a terminal state (completed, failed, cancelled), the
        record is deleted from the database.
      responses:
        "204":
          description: Deleted (terminal execution)
        "200":
          description: Cancelled (was running)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Execution"

  # ── Credentials ──────────────────────────────────────────────────────────────
  /credentials:
    get:
      tags: [Credentials]
      summary: List credentials (no secret data returned)
      responses:
        "200":
          description: List of credentials
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Credential"

    post:
      tags: [Credentials]
      summary: Create a credential
      description: |
        Secret values in `data` are encrypted with AES-256-GCM before storage.
        They are never returned by the list or get endpoints.

        **Credential kinds and expected `data` fields:**

        | kind | fields |
        |------|--------|
        | `postgres` | `connectionString` |
        | `smtp` | `host`, `port`, `user`, `pass`, `from`, `secure` |
        | `sendgrid` | `apiKey`, `from` |
        | `mqtt` | `brokerUrl`, `username`, `password` |
        | `http_bearer` | `token` |
        | `http_basic` | `username`, `password` |
        | `http_api_key` | `headerName`, `headerValue` |
        | `generic` | any key-value pairs |
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, kind, data]
              properties:
                name:
                  type: string
                  example: "Production DB"
                kind:
                  $ref: "#/components/schemas/CredentialKind"
                data:
                  type: object
                  additionalProperties:
                    type: string
                  example: { connectionString: "postgresql://user:pass@host/db" }
      responses:
        "201":
          description: Credential created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Credential"
  /credentials/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid

    put:
      tags: [Credentials]
      summary: Update a credential
      description: Provide `name` to rename, `data` to re-encrypt with new values.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                data:
                  type: object
                  additionalProperties:
                    type: string
      responses:
        "200":
          description: Updated credential
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Credential"
        "404":
          description: Not found

    delete:
      tags: [Credentials]
      summary: Delete a credential
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found

  # ── Webhooks ─────────────────────────────────────────────────────────────────
...